- init refactor
