--- .gitignore
@@ -1,9 +1,0 @@
-.idea/
-composer.lock
-vendor/
-clover.xml
-coveralls.json
-tests/server.exe
-tests/server
-**/vendor
-vendor_php
