--- composer.json
@@ -1,40 +1,0 @@
-{
-  "name": "spiral/goridge",
-  "type": "goridge",
-  "description": "High-performance PHP-to-Golang RPC bridge",
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "Anton Titov / Wolfy-J",
-      "email": "wolfy.jd@gmail.com"
-    }
-  ],
-  "require": {
-    "php": ">=7.2",
-    "ext-json": "*"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "~8.0",
-    "spiral/code-style": "^1.0",
-    "phpstan/phpstan": "^0.12.23"
-  },
-  "autoload": {
-    "psr-4": {
-      "Spiral\\Goridge\\": "src/"
-    },
-    "files": [
-      "src/functions.php"
-    ]
-  },
-  "config": {
-    "vendor-dir": "vendor_php"
-  },
-  "scripts": {
-    "analyze": "phpstan analyze -c ./phpstan.neon.dist --no-progress --ansi"
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Spiral\\Tests\\": "tests/Cases/"
-    }
-  }
-}
