--- constants.go
@@ -1,3 +1,0 @@
-package goridge
-
-const BufferSize = 10485760 // 10 Mb
