--- examples/composer.json
@@ -1,5 +1,0 @@
-{
-  "require": {
-    "spiral/goridge": "^2.0.0"
-  }
-}
