+++ socket.go
@@ -17,7 +17,7 @@ func NewSocketRelay(rwc io.ReadWriteCloser) *SocketRelay {
 }
 
 // Send signed (prefixed) data to PHP process.
-func (rl *SocketRelay) Send(data []byte, flags byte) (err error) {
+func (rl *SocketRelay) Send(f Frame) (err error) {
 	prefix := NewPrefix().WithFlags(flags).WithSize(uint64(len(data)))
 	if _, err := rl.rwc.Write(prefix[:]); err != nil {
 		return err
@@ -31,7 +31,7 @@ func (rl *SocketRelay) Send(data []byte, flags byte) (err error) {
 }
 
 // Receive data from the underlying process and returns associated prefix or error.
-func (rl *SocketRelay) Receive() (data []byte, p Prefix, err error) {
+func (rl *SocketRelay) Receive() (f Frame, err error) {
 	defer func() {
 		if rErr, ok := recover().(error); ok {
 			err = rErr
