--- src/Exceptions/RelayFactoryException.php
@@ -1,15 +1,0 @@
-<?php
-
-/**
- * Dead simple, high performance, drop-in bridge to Golang RPC with zero dependencies
- *
- * @author Valentin V
- */
-
-declare(strict_types=1);
-
-namespace Spiral\Goridge\Exceptions;
-
-class RelayFactoryException extends GoridgeException
-{
-}
