--- src/Exceptions/TransportException.php
@@ -1,15 +1,0 @@
-<?php
-
-/**
- * Dead simple, high performance, drop-in bridge to Golang RPC with zero dependencies
- *
- * @author Wolfy-J
- */
-
-declare(strict_types=1);
-
-namespace Spiral\Goridge\Exceptions;
-
-class TransportException extends RelayException
-{
-}
