--- src/StringableRelayInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Spiral\Goridge;
-
-interface StringableRelayInterface
-{
-    /**
-     * @return string
-     */
-    public function __toString(): string;
-}
