--- tests/Cases/Legacy/UnixSocketTestIgnored.php
@@ -1,22 +1,0 @@
-<?php
-
-/**
- * Dead simple, high performance, drop-in bridge to Golang RPC with zero dependencies
- *
- * @author Wolfy-J
- */
-
-declare(strict_types=1);
-
-namespace Spiral\Tests\Legacy;
-
-use Spiral\Goridge\RPC;
-use Spiral\Goridge\SocketRelay;
-
-class UnixSocketTestIgnored extends \Spiral\Tests\UnixSocketTestIgnored
-{
-    protected function makeRPC(): RPC
-    {
-        return new RPC(new Relay(new SocketRelay(static::SOCK_ADDR, static::SOCK_PORT, static::SOCK_TYPE)));
-    }
-}
