--- tests/Cases/TCPSocketTest.php
@@ -1,20 +1,0 @@
-<?php
-
-/**
- * Dead simple, high performance, drop-in bridge to Golang RPC with zero dependencies
- *
- * @author Wolfy-J
- */
-
-declare(strict_types=1);
-
-namespace Spiral\Tests;
-
-use Spiral\Goridge\SocketRelay;
-
-class TCPSocketTest extends RPCTest
-{
-    public const SOCK_ADDR = '127.0.0.1';
-    public const SOCK_PORT = 7079;
-    public const SOCK_TYPE = SocketRelay::SOCK_TCP;
-}
