--- tests/docker-compose.yml
@@ -1,10 +1,0 @@
-version: '3'
-
-services:
-  goridge:
-    container_name: goridge
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports: ["7079:7079"]
-    command: ["/app/server"]
