--- tests/go.mod
@@ -1,5 +1,0 @@
-module test_server
-
-go 1.14
-
-require github.com/spiral/goridge/v2 v2.3.3
